Split share extension into three files
- ShareViewController.swift: Platform view controllers (iOS/macOS)
- ShareURLHandler.swift: URL parsing, routing, and helper functions
- UnsupportedURLView.swift: SwiftUI view for unsupported URLs
